Fixed multiple problems with the packet resend logic

+++ src/packet.rs
@@ -19,7 +19,7 @@ pub type StreamNumber = u8;
 /// The fragment number of a fragmented packet
 pub type FragmentNumber = u8;
 /// The sequence number of a packet
-pub type SequenceNumber = u32;
+pub type SequenceNumber = u16;
 pub type AckFieldNumber = u32;
 pub type ConnectionId = u16;
